Refactor Codes : Unify the Response to one file.
