+++ HabitBread/app/src/main/java/com/example/habitbread/adapter/RankListAdapter.kt
@@ -7,7 +7,6 @@ import android.view.ViewGroup
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import com.example.habitbread.R
-import com.example.habitbread.data.Ranking
 import kotlinx.android.synthetic.main.item_rank.view.*
 
 class RankListAdapter(private val context: Context?) : RecyclerView.Adapter<RankListAdapter.RankingListViewHolder>() {
