+++ HabitBread/app/src/main/java/com/example/habitbread/data/BaseResponse.kt
@@ -40,3 +40,36 @@ data class BaseResponse(
     @SerializedName("message")
     val message: String
 )
+
+data class NewHabit(
+    @SerializedName("title")
+    val title: String,
+    @SerializedName("category")
+    val category: String,
+    @SerializedName("dayOfWeek")
+    val dayOfWeek: String,
+    @SerializedName("alarmTime")
+    val alarmTime: String?
+)
+
+data class Ranking(
+    @SerializedName("userId")
+    val userId : Int,
+    @SerializedName("userName")
+    val userName : String,
+    @SerializedName("exp")
+    val exp: Int,
+    @SerializedName("achievement")
+    val achievement : Int,
+    @SerializedName("rank")
+    val rank : String
+)
+
+data class RankResponse(
+    @SerializedName("user")
+    val user : Ranking,
+    @SerializedName("userTotalCount")
+    val userTotalCount: Int,
+    @SerializedName("rankings")
+    val rankings : List<Ranking>
+)
