--- HabitBread/app/src/main/java/com/example/habitbread/data/NewHabit.kt
@@ -1,8 +1,0 @@
-package com.example.habitbread.data
-
-data class NewHabit (
-    val title: String,
-    val category: String,
-    val dayOfWeek: String,
-    val alarmTime: String?
-)
