--- HabitBread/app/src/main/java/com/example/habitbread/data/Ranking.kt
@@ -1,16 +1,0 @@
-package com.example.habitbread.data
-
-import com.google.gson.annotations.SerializedName
-
-data class Ranking(
-    @SerializedName("userId")
-    val userId : Int,
-    @SerializedName("userName")
-    val userName : String,
-    @SerializedName("exp")
-    val exp: Int,
-    @SerializedName("achievement")
-    val achievement : Int,
-    @SerializedName("rank")
-    val rank : String
-)
