+++ HabitBread/app/src/main/java/com/example/habitbread/interface/RankHandler.kt
@@ -1,7 +1,5 @@
 package com.example.habitbread.`interface`
 
-import com.example.habitbread.data.Ranking
-
 interface RankHandler {
     fun onResult(myRank: Ranking, totalCount: Int, allRanks : List<Ranking>)
 }
